+++ Source/ArticyEditor/Private/ArticyEditorModule.cpp
@@ -61,7 +61,7 @@ void FArticyEditorModule::ShutdownModule()
 {
 	if (UObjectInitialized())
 	{
-		UnregisterDefaultArticyRefWidgetExtensions();
+		GetCustomizationManager()->Shutdown();
 		UnregisterPluginSettings();
 		
 		if(ConsoleCommands != nullptr)
@@ -191,15 +191,6 @@ void FArticyEditorModule::RegisterPluginSettings() const
 	}	
 }
 
-void FArticyEditorModule::UnregisterDefaultArticyRefWidgetExtensions() const
-{
-	for(const IArticyIdPropertyWidgetCustomizationFactory* DefaultFactory : DefaultArticyRefWidgetCustomizationFactories)
-	{
-		GetCustomizationManager()->UnregisterArticyIdPropertyWidgetCustomizationFactory(DefaultFactory);
-	}
-}
-
-
 void FArticyEditorModule::UnregisterPluginSettings() const
 {
 	if (ISettingsModule* SettingsModule = FModuleManager::GetModulePtr<ISettingsModule>("Settings"))
@@ -226,12 +217,22 @@ void FArticyEditorModule::QueueImport()
 
 void FArticyEditorModule::OpenArticyWindow()
 {
+    // @TODO Engine versioning
+#if ENGINE_MINOR_VERSION < 26
 	FGlobalTabmanager::Get()->InvokeTab(ArticyWindowTabID);
+#else
+	FGlobalTabmanager::Get()->TryInvokeTab(ArticyWindowTabID);
+#endif
 }
 
 void FArticyEditorModule::OpenArticyGVDebugger()
 {
+	// @TODO Engine versioning
+#if ENGINE_MINOR_VERSION < 26
 	FGlobalTabmanager::Get()->InvokeTab(ArticyGVDebuggerTabID);
+#else
+	FGlobalTabmanager::Get()->TryInvokeTab(ArticyGVDebuggerTabID);
+#endif
 }
 
 EImportStatusValidity FArticyEditorModule::CheckImportStatusValidity() const
@@ -255,7 +256,7 @@ EImportStatusValidity FArticyEditorModule::CheckImportStatusValidity() const
 	}
 	
 	TArray<FAssetData> ArticyAssets;
-	AssetRegistryModule.Get().GetAssetsByPath(FName(*ArticyHelpers::ArticyGeneratedFolder), ArticyAssets, true);
+	AssetRegistryModule.Get().GetAssetsByPath(FName(*ArticyHelpers::GetArticyGeneratedFolder()), ArticyAssets, true);
 
 	// check if all assets are actually valid (classes not found would result in a nullptr)
 	for(FAssetData& Data : ArticyAssets)