+++ Source/ArticyEditor/Private/CodeGeneration/ObjectDefinitionsGenerator.cpp
@@ -1,6 +1,5 @@
 //  
 // Copyright (c) articy Software GmbH & Co. KG. All rights reserved.  
- 
 //