+++ Source/ArticyEditor/Private/Customizations/ArticyEditorCustomizationManager.cpp
@@ -40,3 +40,8 @@ void FArticyEditorCustomizationManager::CreateArticyIdPropertyWidgetCustomizatio
 		}
 	}
 }
+
+void FArticyEditorCustomizationManager::Shutdown()
+{
+	ArticyIdPropertyWidgetCustomizationFactories.Empty();
+}