+++ Source/ArticyEditor/Public/Slate/AssetPicker/SArticyObjectAssetPicker.h
@@ -1,4 +1,3 @@
-
 //  
 // Copyright (c) articy Software GmbH & Co. KG. All rights reserved.  
 //
@@ -90,7 +89,7 @@ private:// Widgets
 	TSharedPtr<SComboButton> ClassFilterButton;
 	
 private: // Internal Data
-	TSharedPtr<FAssetFilterCollectionType> FrontendFilters;
+	TSharedPtr<FArticyObjectFilterCollectionType> FrontendFilters;
 	TSharedPtr<FArticyClassRestrictionFilter> ClassFilter;
 	TSharedPtr<FFrontendFilter_ArticyObject> ArticyObjectFilter;