+++ Source/ArticyRuntime/Private/ArticyAsset.cpp
@@ -26,7 +26,7 @@ UObject* UArticyAsset::LoadAsset() const
 		const auto filename = FPaths::GetBaseFilename(AssetRef); //without extension
 
 		//construct the asset path like UE4 wants it
-		auto path = ArticyHelpers::ArticyAssetsFolder / folder / filename;
+		auto path = ArticyHelpers::GetArticyResourcesFolder() / folder / filename;
 		//UE_LOG(LogTemp, Warning, TEXT("Asset Path %s"), *path)
 #if ENGINE_MINOR_VERSION <= 24
 		Asset = ConstructorHelpersInternal::FindOrLoadObject<UObject>(path);